requirements.txt
fastapi
uvicorn
pydantic
prisma==0.15.0
strawberry-graphql
aiobotocore
python-multipart